refactor(architecture): consolidate and clean up abnormality detection and quality control modules
- Removed duplicate implementations of `CleanAbnormalityDetector` and `CleanQualityController`, streamlining the codebase to adhere to Clean Architecture principles.
- Eliminated backward compatibility shims and legacy parameters, ensuring that the domain layer has no dependencies on infrastructure or application layers.
- Updated the `detector.py` and `controller.py` modules to follow proper dependency injection practices, enhancing maintainability and clarity.
- This refactoring addresses critical architectural violations and sets the foundation for improved code integrity and future development.
