+++ ARCHITECTURE_TRUTH.md
@@ -1,0 +1,164 @@
+# 🔍 THE ACTUAL TRUTH ABOUT THE ARCHITECTURE
+
+## What Actually Happened During "Refactoring"
+
+### They Created THREE Versions of Everything
+1. **Original Version** (`detector.py`) - Still exists
+2. **"Clean" Version** (`detector_clean.py`) - With backward compat hacks
+3. **"Pure" Version** (`detector_pure.py`) - Another attempt
+
+### The "Backward Compatibility" Excuse
+They violated Clean Architecture "for backward compatibility":
+```python
+# In domain layer (VIOLATION!)
+if model is None:
+    from brain_go_brrr.infra.adapters.model_adapter import EEGPTModelAdapter
+    model = EEGPTModelAdapter()  # Domain creating infrastructure!
+```
+
+## Directory Types
+
+### REAL Code Directories
+- `api/` - FastAPI endpoints (proper DI mostly)
+- `application/` - Use cases and orchestration
+- `domain/` - Business logic (CONTAMINATED with infra imports)
+- `infra/` - Adapters and external services
+- `core/` - Has 300+ lines of code that should be in domain
+- `utils/` - Actual utilities
+
+### SHIM Directories (Redirects)
+- `config/` → `application/config`
+- `models/` → `infra/ml_models`
+- `preprocessing/` → `domain/preprocessing`
+- `services/` → various locations
+- `visualization/` → `infra/visualization`
+
+## The Real Problems
+
+### 1. Domain Layer is Contaminated
+- **6 files** import from infrastructure
+- **1 file** imports from core
+- Domain creates its own infrastructure objects
+
+### 2. Triple Implementation Pattern
+Instead of refactoring, they created:
+- Original implementation
+- "Clean" implementation (with violations)
+- "Pure" implementation (another attempt)
+
+All three exist simultaneously!
+
+### 3. Core Layer Has Business Logic
+`core/preprocessing_utils.py` contains:
+- PreprocessingConfig (domain entity)
+- BandpassFilter (domain service)
+- PreprocessingPipeline (domain service)
+
+This should ALL be in domain!
+
+### 4. Backward Compatibility Over Architecture
+Every "Clean" class has this pattern:
+```python
+def __init__(self, dependency=None):
+    if dependency is None:
+        # VIOLATION: Domain creating infrastructure
+        from infra import ConcreteImplementation
+        dependency = ConcreteImplementation()
+```
+
+## What They Should Have Done
+
+### Option A: Proper Refactoring
+1. Move code to correct layers
+2. Update imports everywhere
+3. Delete old code
+4. One breaking change, clean result
+
+### Option B: Gradual Migration
+1. Create new clean modules
+2. Mark old ones deprecated
+3. Migrate consumers one by one
+4. Delete old after migration complete
+
+### What They Actually Did: Option C (Worst)
+1. Keep old code
+2. Add new code with violations
+3. Add shims for compatibility
+4. Result: More complexity, same problems
+
+## Quick Metrics
+
+- **26** Python files in domain/
+- **6** domain files violate Clean Architecture
+- **3** versions of abnormality detector
+- **2** versions of quality controller
+- **2** versions of feature extractor
+- **300** lines in core/ that should be in domain/
+- **5** shim directories for backward compatibility
+
+## The Verdict
+
+**The refactoring made things WORSE because:**
+
+1. **Before**: Messy but consistent monolith
+2. **After**: Messy monolith + duplicates + violations + shims
+
+They added complexity without solving problems.
+
+## What Would Actually Fix This
+
+### Week 1: Emergency Surgery
+```bash
+# 1. Remove infra imports from domain (2 hours)
+grep -r "from brain_go_brrr.infra" src/brain_go_brrr/domain/
+# Fix each file to use dependency injection
+
+# 2. Move core logic to domain (1 hour)
+mv src/brain_go_brrr/core/preprocessing_utils.py \
+   src/brain_go_brrr/domain/preprocessing/core.py
+
+# 3. Delete duplicate implementations (2 hours)
+rm src/brain_go_brrr/domain/abnormal/detector.py
+rm src/brain_go_brrr/domain/abnormal/detector_pure.py
+# Keep only detector_clean.py, rename to detector.py
+```
+
+### Week 2: Clean Up
+```bash
+# 4. Remove shim directories (4 hours)
+rm -rf src/brain_go_brrr/{config,models,preprocessing,services,visualization}
+
+# 5. Update all imports (8 hours)
+# Find and replace old paths with new paths
+
+# 6. Add architecture tests (4 hours)
+# Write tests that fail if domain imports from outer layers
+```
+
+### Week 3: Enforce
+```bash
+# 7. Add import linter (2 hours)
+# Configure to prevent future violations
+
+# 8. Document the REAL architecture (2 hours)
+# Update all docs to reflect actual structure
+```
+
+## Summary
+
+The "100% Clean Architecture COMPLETE" claim is FALSE.
+
+What we have is:
+- Original mess
+- Plus "clean" mess
+- Plus backward compatibility mess
+- = Triple mess
+
+**Estimated effort to ACTUALLY fix: 25-30 hours of focused work**
+
+But they'd need to:
+1. Accept breaking changes
+2. Delete backward compatibility
+3. Commit to the architecture
+
+Otherwise, it's just adding more layers to the mess.
