--- src/brain_go_brrr/domain/abnormal/detector_clean.py
@@ -1,414 +1,0 @@
-"""Clean Architecture EEG Abnormality Detection.
-
-This module follows Clean Architecture principles - the domain layer
-has NO dependencies on infrastructure or application layers.
-All dependencies are inverted through ports/interfaces.
-"""
-
-import time
-from dataclasses import dataclass
-from enum import Enum
-from typing import Any
-
-import numpy as np
-import numpy.typing as npt
-import torch
-
-from brain_go_brrr._typing import MNERaw
-from brain_go_brrr.domain.ports import (
-    AbnormalityConfigPort,
-    EEGModelPort,
-    LoggerPort,
-    PreprocessorPort,
-)
-
-
-class TriageLevel(str, Enum):
-    """Clinical triage levels for EEG prioritization."""
-
-    NORMAL = "NORMAL"  # Low priority
-    ROUTINE = "ROUTINE"  # Standard workflow (< 48 hours)
-    EXPEDITE = "EXPEDITE"  # Priority review (< 4 hours)
-    URGENT = "URGENT"  # Immediate review needed
-
-
-@dataclass
-class AbnormalityResult:
-    """Result of abnormality detection analysis."""
-
-    is_abnormal: bool
-    confidence: float
-    triage_level: TriageLevel
-    processing_time_ms: float
-    features_shape: tuple[int, ...]
-    metadata: dict[str, Any]
-
-
-class CleanAbnormalityDetector:
-    """Clean Architecture Abnormality Detector using dependency injection.
-
-    This class follows Clean Architecture principles:
-    - Domain logic is pure (no infrastructure dependencies)
-    - All dependencies are injected through ports/interfaces
-    - Business rules are isolated from implementation details
-    """
-
-    def __init__(
-        self,
-        model: EEGModelPort | None = None,
-        preprocessor: PreprocessorPort | None = None,
-        config: AbnormalityConfigPort | None = None,
-        logger: LoggerPort | None = None,
-        linear_probe: torch.nn.Module | None = None,
-        # Legacy parameters for backward compatibility
-        model_path: Any = None,
-        device: str = "cpu",
-        **_ignored: Any,
-    ):
-        """Initialize detector with injected dependencies.
-
-        Args:
-            model: EEG model for feature extraction (port, optional for back-compat)
-            preprocessor: EEG preprocessor (port, optional for back-compat)
-            config: Configuration (port, optional for back-compat)
-            logger: Logger (port, optional)
-            linear_probe: Linear probe head for classification (optional)
-            model_path: Legacy parameter (for backward compatibility)
-            device: Legacy parameter (for backward compatibility)
-            **_ignored: Other legacy parameters (ignored)
-        """
-        # Create defaults if not provided (for backward compatibility)
-        if model is None:
-            from brain_go_brrr.infra.adapters.model_adapter import EEGPTModelAdapter
-
-            model = EEGPTModelAdapter(
-                model_path=str(model_path) if model_path else "data/models/pretrained/eegpt.ckpt",
-                device=device,
-            )
-        if preprocessor is None:
-            from brain_go_brrr.infra.adapters.model_adapter import EEGPreprocessorAdapter
-
-            preprocessor = EEGPreprocessorAdapter()
-        if config is None:
-            # Create minimal config adapter
-            class MinimalModel:
-                feature_dim: int = 512  # EEGPT's actual embedding dimension
-
-            class MinimalConfig:
-                def __init__(self) -> None:
-                    self.model = MinimalModel()
-                    self.confidence_threshold = 0.5  # Add as attribute too
-                    self.channels: list[str] = []  # Empty list for minimal config
-
-                def get_confidence_threshold(self) -> float:
-                    return self.confidence_threshold
-
-                def get_min_confidence(self) -> float:
-                    return 0.3
-
-                def get_required_channels(self) -> list[str]:
-                    return self.channels
-
-                def get_bandpass_low(self) -> float:
-                    return 0.5
-
-                def get_bandpass_high(self) -> float:
-                    return 50.0
-
-            config = MinimalConfig()  # type: ignore[assignment]
-
-        self.model = model
-        self.preprocessor = preprocessor
-        self.config = config
-        self.logger = logger
-        self.linear_probe = linear_probe
-        # Use the provided device, or auto-detect if cuda requested but not available
-        if device == "cuda" and not torch.cuda.is_available():
-            self.device = torch.device("cpu")
-        else:
-            self.device = torch.device(device)
-
-        # Store legacy parameters
-        self.model_path = model_path
-
-        # Set feature_dim from config if available, else default to 512
-        # The test expects 512 as default (EEGPT's actual embedding dim)
-        if config and hasattr(config, "model") and hasattr(config.model, "feature_dim"):
-            self.feature_dim = config.model.feature_dim
-        else:
-            self.feature_dim = 512  # EEGPT's actual embedding dimension
-
-        # Initialize linear probe if not provided
-        if self.linear_probe is None:
-            self._initialize_linear_probe()
-
-        # Backward compatibility: expose linear_probe as classifier
-        self.classifier = self.linear_probe
-
-    def _init_model(self) -> None:
-        """Initialize model (backward compatibility method)."""
-        # This method exists for backward compatibility with tests
-        pass
-
-    def _initialize_linear_probe(self) -> None:
-        """Initialize linear probe head for binary classification."""
-        # Use the configured feature_dim
-        feature_dim = self.feature_dim
-
-        self.linear_probe = torch.nn.Sequential(
-            torch.nn.Linear(feature_dim, 256),
-            torch.nn.ReLU(),
-            torch.nn.Dropout(0.5),
-            torch.nn.Linear(256, 64),
-            torch.nn.ReLU(),
-            torch.nn.Dropout(0.3),
-            torch.nn.Linear(64, 2),  # Binary classification
-        )
-        # Only move to device if not CPU (to avoid cuda issues in tests)
-        if self.device.type != "cpu":
-            self.linear_probe = self.linear_probe.to(self.device)
-
-        if self.logger:
-            self.logger.info(f"Initialized linear probe with {feature_dim} input features")
-
-    def detect(self, raw: MNERaw) -> AbnormalityResult:
-        """Detect abnormalities in EEG recording.
-
-        This is the core domain logic - pure business rules without
-        any infrastructure concerns.
-
-        Args:
-            raw: Raw EEG data
-
-        Returns:
-            AbnormalityResult with detection outcome
-        """
-        start_time = time.perf_counter()
-
-        # Validate model compatibility before processing
-        try:
-            self.validate_model_compatibility()
-        except RuntimeError as e:
-            if "dimension mismatch" in str(e):
-                # Re-raise with more context for the test
-                raise RuntimeError("Model/classifier dimension mismatch") from e
-            raise
-
-        # Step 1: Preprocess the EEG data
-        preprocessed = self.preprocessor.preprocess(
-            raw,
-            bandpass=(0.5, 45.0),  # Standard EEG bandpass
-            notch=50.0,  # Standard power line frequency
-        )
-
-        # Step 2: Convert to array for model input
-        eeg_array = self.preprocessor.transform_to_array(preprocessed)
-
-        # Step 3: Extract features using the model
-        features = self.model.extract_features(
-            eeg_array,
-            sampling_rate=int(preprocessed.info["sfreq"]),
-        )
-
-        # Step 4: Run inference with linear probe
-        confidence = self._run_inference(features)
-
-        # Step 5: Apply business rules for classification
-        is_abnormal = confidence > self.config.confidence_threshold  # type: ignore[union-attr]
-        triage_level = self._determine_triage_level(confidence, is_abnormal)
-
-        # Step 6: Calculate processing time
-        processing_time_ms = (time.perf_counter() - start_time) * 1000
-
-        # Step 7: Log if logger is available
-        if self.logger:
-            self.logger.info(
-                f"Detection complete: abnormal={is_abnormal}, "
-                f"confidence={confidence:.3f}, triage={triage_level.value}, "
-                f"time={processing_time_ms:.1f}ms"
-            )
-
-        return AbnormalityResult(
-            is_abnormal=is_abnormal,
-            confidence=float(confidence),
-            triage_level=triage_level,
-            processing_time_ms=processing_time_ms,
-            features_shape=features.shape,
-            metadata={
-                "n_channels": eeg_array.shape[0],
-                "n_samples": eeg_array.shape[1],
-                "sampling_rate": preprocessed.info["sfreq"],
-                "threshold": self.config.confidence_threshold,  # type: ignore[union-attr]
-            },
-        )
-
-    def _run_inference(self, features: npt.NDArray[np.float32] | torch.Tensor) -> float:
-        """Run inference using the linear probe.
-
-        Args:
-            features: Extracted features from EEG model
-
-        Returns:
-            Confidence score for abnormality (0-1)
-        """
-        with torch.inference_mode():
-            # Convert to tensor and move to device
-            if isinstance(features, torch.Tensor):
-                features_tensor = features.float().to(self.device)
-            else:
-                features_tensor = torch.from_numpy(features).float().to(self.device)
-
-            # Add batch dimension if needed
-            if features_tensor.dim() == 1:
-                features_tensor = features_tensor.unsqueeze(0)
-
-            # Handle dimension mismatch - if features are 2048 (full EEGPT) but classifier expects 768
-            # we need to either average or truncate
-            if features_tensor.shape[-1] == 2048 and self.linear_probe[0].in_features == 768:  # type: ignore[index]
-                # Reshape to (batch, 4, 512) and average over tokens
-                batch_size = features_tensor.shape[0]
-                features_tensor = features_tensor.view(batch_size, 4, 512).mean(dim=1)
-                # Now we have (batch, 512) but classifier might expect 768
-                # Pad with zeros to match
-                padding = torch.zeros(batch_size, 768 - 512, device=features_tensor.device)
-                features_tensor = torch.cat([features_tensor, padding], dim=1)
-
-            # Run through linear probe if available
-            if self.linear_probe is not None:
-                logits = self.linear_probe(features_tensor)
-                # Apply softmax to get probabilities
-                probs = torch.softmax(logits, dim=-1)
-            else:
-                # Fallback: use heuristic if no linear probe
-                # Use simple heuristic for abnormality score
-                abnormal_score = float(np.mean(features) > 0.5)
-                # Create pseudo-probabilities
-                probs = torch.tensor([[1 - abnormal_score, abnormal_score]])
-
-            # Return abnormality probability (class 1)
-            return probs[0, 1].cpu().item()
-
-    def _determine_triage_level(self, confidence: float, is_abnormal: bool) -> TriageLevel:
-        """Determine clinical triage level based on confidence and abnormality.
-
-        This encapsulates the business rules for clinical prioritization.
-
-        Args:
-            confidence: Confidence score (0-1)
-            is_abnormal: Whether EEG is classified as abnormal
-
-        Returns:
-            Appropriate triage level
-        """
-        if not is_abnormal:
-            return TriageLevel.NORMAL
-
-        # Business rules for triage based on confidence
-        if confidence >= 0.95:
-            return TriageLevel.URGENT  # Very high confidence abnormality
-        elif confidence >= 0.85:
-            return TriageLevel.EXPEDITE  # High confidence
-        elif confidence >= 0.70:
-            return TriageLevel.ROUTINE  # Moderate confidence
-        else:
-            return TriageLevel.NORMAL  # Low confidence, needs review
-
-    def validate_input(self, raw: MNERaw) -> bool:
-        """Validate input EEG data meets requirements.
-
-        Pure domain validation logic.
-
-        Args:
-            raw: Raw EEG data
-
-        Returns:
-            True if valid, raises exception otherwise
-        """
-        # Check sampling rate
-        if raw.info["sfreq"] < 100:
-            raise ValueError(f"Sampling rate too low: {raw.info['sfreq']}Hz (minimum 100Hz)")
-
-        # Check duration
-        duration = raw.n_times / raw.info["sfreq"]
-        if duration < 10:
-            raise ValueError(f"Recording too short: {duration:.1f}s (minimum 10s)")
-
-        # Check channels
-        available_channels = set(raw.ch_names)
-        required_channels = set(self.config.channels)  # type: ignore[union-attr]
-        missing = required_channels - available_channels
-
-        if missing:
-            raise ValueError(f"Missing required channels: {missing}")
-
-        return True
-
-    def validate_model_compatibility(self, _feature_dim: int | None = None) -> None:
-        """Validate model compatibility (for backward compatibility with tests).
-
-        Args:
-            feature_dim: Expected feature dimension (optional)
-
-        Raises:
-            RuntimeError: If dimensions mismatch
-        """
-        # Check if model dimensions match classifier expectations
-        if hasattr(self.model, "embedding_dim"):
-            model_output_dim = self.model.embedding_dim
-
-            # Test expects 512 to be valid (no error), 256 to be invalid (error)
-            # The test sets embedding_dim to 256 to trigger the error case
-            if model_output_dim == 256:
-                # This is the test case for incompatible dimensions
-                raise RuntimeError("dimension mismatch")
-
-            # For the actual detection flow, check against linear probe input
-            if hasattr(self, "linear_probe") and self.linear_probe is not None:
-                expected_dim = self.linear_probe[0].in_features  # type: ignore[index]
-                # Model outputs 512 but classifier expects 768 is OK (we pad)
-                # But if model outputs something completely wrong, error
-                if model_output_dim not in [256, 512, 768, expected_dim]:
-                    raise RuntimeError(
-                        f"Model/classifier dimension mismatch: model outputs {model_output_dim}, classifier expects {expected_dim}"
-                    )
-
-    def detect_abnormality(self, raw: MNERaw) -> dict[str, Any]:
-        """Detect abnormality (backward compatibility method)."""
-        result = self.detect(raw)
-        return {
-            "is_abnormal": result.is_abnormal,
-            "confidence": result.confidence,
-            "triage_level": result.triage_level.value,
-            "processing_time_ms": result.processing_time_ms,
-        }
-
-    def _load_classifier_weights(self, path: Any) -> None:
-        """Load classifier weights (backward compatibility method)."""
-        # Load weights and validate dimensions
-        state_dict = torch.load(path) if not isinstance(path, dict) else path
-
-        # Check first layer dimensions against the linear probe
-        if "0.weight" in state_dict and self.linear_probe is not None:
-            weight_shape = state_dict["0.weight"].shape
-            expected_dim = self.linear_probe[0].in_features  # type: ignore[index]  # Get from actual classifier
-            actual_dim = weight_shape[1]  # Input dimension is second dim of weight matrix
-
-            if actual_dim != expected_dim:
-                raise RuntimeError(
-                    f"Classifier dimension mismatch: expected {expected_dim}, got {actual_dim}"
-                )
-
-    def _predict_window(self, window: npt.NDArray[np.float32]) -> float:
-        """Predict abnormality score for a single window (backward compatibility).
-
-        Args:
-            window: EEG window data (channels, samples)
-
-        Returns:
-            Abnormality score (0-1)
-        """
-        # Extract features from window
-        features = self.model.extract_features(window, sampling_rate=256)
-
-        # Run inference
-        return self._run_inference(features)
